collect factory methods together
